[ion] millis() and micros() returns uint32_t and are defined in their own header

+++ ion/src/blackbox/ion.cpp
@@ -8,12 +8,12 @@ void Ion::msleep(long ms) {
 
 static auto start = std::chrono::high_resolution_clock::now();
 
-long Ion::millis() {
+uint32_t Ion::millis() {
   auto elapsed = std::chrono::high_resolution_clock::now() - start;
   return std::chrono::duration_cast<std::chrono::milliseconds>(elapsed).count();
 }
 
-long Ion::micros() {
+uint32_t Ion::micros() {
   auto elapsed = std::chrono::high_resolution_clock::now() - start;
   return std::chrono::duration_cast<std::chrono::microseconds>(elapsed).count();
 }
